wip: feat: general additions and fixes for trucks and fleets

+++ src/main/java/com/empacoters/antsback/logistics/application/usecases/DeleteTruckUseCase.java
@@ -1,6 +1,7 @@
 package com.empacoters.antsback.logistics.application.usecases;
 
 import com.empacoters.antsback.logistics.domain.repository.TruckRepository;
+import com.empacoters.antsback.shared.exception.NotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -13,6 +14,11 @@ public class DeleteTruckUseCase {
     }
     public void execute(Long id)
     {
+        var truck = truckRepository.byId(id);
+
+        if (truck == null) {
+            throw new NotFoundException("O caminhão solicitado não existe");
+        }
         truckRepository.delete(id);
     }
